+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 0.0.2 (TBA)
 * Improvement: Added unit tests
+* Improvement: Added subscribe / unsubsbscribe methods for candles and ticker to the ticketmanager
 * Bugfix: The ticks now a volume of 0
 
 # Version 0.0.1 (27.12.2017)